requirements.txt
astroid==2.2.5
certifi==2020.4.5.1
colorama==0.4.1
Django==2.2.12
djangorestframework==3.11.0
isort==4.3.20
joblib==0.14.1
lazy-object-proxy==1.4.1
mccabe==0.6.1
numpy==1.18.4
pylint==2.3.1
pytz==2020.1
scikit-learn==0.23.0
scipy==1.4.1
six==1.12.0
sklearn==0.0
sqlparse==0.3.1
threadpoolctl==2.0.0
typed-ast==1.3.5
wincertstore==0.2
wrapt==1.11.1
